shift parseUtil as suggested
