+++ src/test/java/seedu/todo/controllers/DateParserTest.java
@@ -1,4 +1,4 @@
-package seedu.todo.commons.util;
+package seedu.todo.controllers;
 
 import static org.junit.Assert.*;
 
@@ -7,8 +7,10 @@ import java.util.Map;
 
 import org.junit.Test;
 
+import seedu.todo.controllers.concerns.DateParser;
+
 //@@author A0139922Y
-public class ParseUtilTest {
+public class DateParserTest {
     
     Map<String, String[]> parsedResult = new HashMap<String, String[]>();
     public static final String EMPTY_TOKEN = "";
@@ -38,48 +40,48 @@ public class ParseUtilTest {
     @Test
     public void testIsTokenNull_true() {
         parsedResult.put(TEST_TOKEN, TOKEN_RESULT);
-        assertTrue(ParseUtil.isTokenNull(parsedResult, EMPTY_TOKEN));
+        assertTrue(DateParser.isTokenNull(parsedResult, EMPTY_TOKEN));
         
         parsedResult.put(TEST_TOKEN, NULL_TOKEN_RESULT);
-        assertTrue(ParseUtil.isTokenNull(parsedResult, TEST_TOKEN));
+        assertTrue(DateParser.isTokenNull(parsedResult, TEST_TOKEN));
     }
     
     @Test
     public void testIsTokenNull_false() {
         parsedResult.put(TEST_TOKEN, TOKEN_RESULT);
-        assertFalse(ParseUtil.isTokenNull(parsedResult, TEST_TOKEN));
+        assertFalse(DateParser.isTokenNull(parsedResult, TEST_TOKEN));
     }
 
     @Test
     public void testDoesTokenContainKeyword_true() {
         parsedResult.put(TEST_TOKEN, TOKEN_RESULT);
-        assertTrue(ParseUtil.doesTokenContainKeyword(parsedResult, TEST_TOKEN, TOKEN_RESULT[TOKEN_RESULT_INDEX_ONE]));
+        assertTrue(DateParser.doesTokenContainKeyword(parsedResult, TEST_TOKEN, TOKEN_RESULT[TOKEN_RESULT_INDEX_ONE]));
     }
     
     @Test
     public void testDoesTokenContainKeyword_false() {
         parsedResult.put(TEST_TOKEN, TOKEN_RESULT);
-        assertFalse(ParseUtil.doesTokenContainKeyword(parsedResult, TEST_TOKEN, TOKEN_KEYWORD_DOES_NOT_EXIST));
-        assertFalse(ParseUtil.doesTokenContainKeyword(parsedResult, EMPTY_TOKEN, TOKEN_RESULT[TOKEN_RESULT_INDEX_ONE]));
-        assertFalse(ParseUtil.doesTokenContainKeyword(parsedResult, EMPTY_TOKEN, NULL_TOKEN_RESULT[TOKEN_RESULT_INDEX_ONE]));
+        assertFalse(DateParser.doesTokenContainKeyword(parsedResult, TEST_TOKEN, TOKEN_KEYWORD_DOES_NOT_EXIST));
+        assertFalse(DateParser.doesTokenContainKeyword(parsedResult, EMPTY_TOKEN, TOKEN_RESULT[TOKEN_RESULT_INDEX_ONE]));
+        assertFalse(DateParser.doesTokenContainKeyword(parsedResult, EMPTY_TOKEN, NULL_TOKEN_RESULT[TOKEN_RESULT_INDEX_ONE]));
     }
 
     @Test
     public void testGetTokenResult_null_test() {
-        assertNotNull(ParseUtil.getTokenResult(parsedResult, TEST_TOKEN), TOKEN_RESULT[TOKEN_RESULT_INDEX_TWO]);
-        assertNull(ParseUtil.getTokenResult(parsedResult, EMPTY_TOKEN));
+        assertNotNull(DateParser.getTokenResult(parsedResult, TEST_TOKEN), TOKEN_RESULT[TOKEN_RESULT_INDEX_TWO]);
+        assertNull(DateParser.getTokenResult(parsedResult, EMPTY_TOKEN));
     }
     
     @Test
     public void testGetTokenResult() {
         parsedResult.put(TEST_TOKEN, TOKEN_RESULT);
-        assertEquals(ParseUtil.getTokenResult(parsedResult, TEST_TOKEN), TOKEN_RESULT[TOKEN_RESULT_INDEX_TWO]);
+        assertEquals(DateParser.getTokenResult(parsedResult, TEST_TOKEN), TOKEN_RESULT[TOKEN_RESULT_INDEX_TWO]);
     }
     
     @Test
     public void testGetTokenResult_not_equals() {
         parsedResult.put(TEST_TOKEN, TOKEN_RESULT);
-        assertNotEquals(ParseUtil.getTokenResult(parsedResult, TEST_TOKEN), TOKEN_RESULT[TOKEN_RESULT_INDEX_ONE]);
+        assertNotEquals(DateParser.getTokenResult(parsedResult, TEST_TOKEN), TOKEN_RESULT[TOKEN_RESULT_INDEX_ONE]);
     }
 
     @Test
@@ -87,11 +89,11 @@ public class ParseUtilTest {
         parsedResult.put(date_on_format, date_on_result);
         parsedResult.put(date_from_format, incorrect_date_result);
         parsedResult.put(date_to_format, incorrect_date_result);
-        assertNotNull(ParseUtil.parseDates(parsedResult));
+        assertNotNull(DateParser.parseDates(parsedResult));
         
         parsedResult.put(date_on_format, incorrect_date_result);
         parsedResult.put(date_to_format, date_to_result);
-        assertNotNull(ParseUtil.parseDates(parsedResult));
+        assertNotNull(DateParser.parseDates(parsedResult));
     }
     
     @Test
@@ -99,7 +101,7 @@ public class ParseUtilTest {
         parsedResult.put(date_on_format, incorrect_date_result);
         parsedResult.put(date_from_format, incorrect_date_result);
         parsedResult.put(date_to_format, incorrect_date_result);
-        assertNull(ParseUtil.parseDates(parsedResult)); 
+        assertNull(DateParser.parseDates(parsedResult)); 
     }
     
     @Test
@@ -107,7 +109,7 @@ public class ParseUtilTest {
         parsedResult.put(date_on_format, date_on_result);
         parsedResult.put(date_from_format, date_from_result);
         parsedResult.put(date_to_format, date_to_result);
-        assertArrayEquals(test_result, ParseUtil.parseDates(parsedResult));
+        assertArrayEquals(test_result, DateParser.parseDates(parsedResult));
     }
     
     @Test
@@ -115,6 +117,6 @@ public class ParseUtilTest {
         parsedResult.put(date_on_format, date_on_result);
         parsedResult.put(date_from_format, date_from_result);
         parsedResult.put(date_to_format, date_to_result);
-        assertNotEquals(incorrect_test_result, ParseUtil.parseDates(parsedResult));
+        assertNotEquals(incorrect_test_result, DateParser.parseDates(parsedResult));
     }
 }
